# .migration-backup/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * O'price Ng marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Listing {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  sellerId: number;
  sellerName: string;
  sellerUsername: string;
  /** @nullable */
  sellerAvatar: string | null;
  isVerifiedSeller: boolean;
  condition: string;
  category: string;
  location: string;
  /** @nullable */
  shippingInfo?: string | null;
  images?: string[];
  viewCount: number;
  watchCount: number;
  offerCount: number;
  isAuction: boolean;
  /** @nullable */
  auctionEndsAt?: string | null;
  /** @nullable */
  currentBid?: number | null;
  isSponsored: boolean;
  isWatched?: boolean;
  isBookmarked?: boolean;
  createdAt: string;
}

export interface ListingInput {
  title: string;
  description?: string;
  price: number;
  condition: string;
  category: string;
  location: string;
  shippingInfo?: string;
  images?: string[];
  isAuction?: boolean;
}

export interface ListingUpdate {
  title?: string;
  description?: string;
  price?: number;
  condition?: string;
  location?: string;
  shippingInfo?: string;
}

export interface ListingsPage {
  listings: Listing[];
  total: number;
  page: number;
  limit: number;
}

export interface UserProfile {
  id: number;
  name: string;
  username: string;
  /** @nullable */
  avatar: string | null;
  /** @nullable */
  coverImage?: string | null;
  /** @nullable */
  bio: string | null;
  location: string;
  joinDate: string;
  rating: number;
  totalSales: number;
  totalListings: number;
  isVerified: boolean;
}

export interface Offer {
  id: number;
  listingId: number;
  buyerName: string;
  buyerUsername: string;
  amount: number;
  status: string;
  /** @nullable */
  message?: string | null;
  createdAt: string;
}

export interface OfferInput {
  amount: number;
  message?: string;
}

export interface Bid {
  id: number;
  listingId: number;
  bidderName: string;
  bidderUsername: string;
  amount: number;
  createdAt: string;
}

export interface BidInput {
  amount: number;
}

export interface WatchResult {
  isWatched: boolean;
  watchCount: number;
}

export interface BookmarkResult {
  isBookmarked: boolean;
}

export interface Conversation {
  id: number;
  otherPartyName: string;
  otherPartyUsername: string;
  /** @nullable */
  otherPartyAvatar: string | null;
  listingTitle: string;
  /** @nullable */
  listingImage?: string | null;
  lastMessage: string;
  lastMessageAt: string;
  unreadCount: number;
}

export interface Message {
  id: number;
  conversationId: number;
  senderUsername: string;
  content: string;
  createdAt: string;
  isMine: boolean;
}

export interface MessageInput {
  content: string;
}

export interface Notification {
  id: number;
  type: string;
  text: string;
  /** @nullable */
  actorName?: string | null;
  /** @nullable */
  actorUsername?: string | null;
  /** @nullable */
  actorAvatar?: string | null;
  /** @nullable */
  listingTitle?: string | null;
  /** @nullable */
  listingImage?: string | null;
  isRead: boolean;
  createdAt: string;
}

export interface FeedStats {
  totalListings: number;
  activeListings: number;
  totalSellers: number;
  totalCategories: number;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  listingCount: number;
  /** @nullable */
  icon?: string | null;
}

export interface TrendingSearch {
  term: string;
  count: number;
}

export type GetListingsParams = {
page?: number;
limit?: number;
category?: string;
search?: string;
condition?: string;
};