# .migration-backup/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: O'price Ng marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /listings/{id}/offers:
    get:
      operationId: getListingOffers
      tags: [offers]
      summary: Get offers on a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Offers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Offer"
    post:
      operationId: createOffer
      tags: [offers]
      summary: Make an offer on a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfferInput"
      responses:
        "201":
          description: Offer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
  /listings/{id}/bid:
    post:
      operationId: placeBid
      tags: [listings]
      summary: Place an auction bid
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BidInput"
      responses:
        "201":
          description: Bid placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bid"
...
components:
  schemas:
...
    Offer:
      type: object
      required: [id, listingId, buyerName, buyerUsername, amount, status, createdAt]
      properties:
        id:
          type: integer
        listingId:
          type: integer
        buyerName:
          type: string
        buyerUsername:
          type: string
        amount:
          type: number
        status:
          type: string
        message:
          type: ["string", "null"]
        createdAt:
          type: string
    OfferInput:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
        message:
          type: string
    Bid:
      type: object
      required: [id, listingId, bidderName, bidderUsername, amount, createdAt]
      properties:
        id:
          type: integer
        listingId:
          type: integer
        bidderName:
          type: string
        bidderUsername:
          type: string
        amount:
          type: number
        createdAt:
          type: string
    BidInput:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
...