# .migration-backup/lib/api-zod/src/generated/types/bid.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * O'price Ng marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface Bid {
  id: number;
  listingId: number;
  bidderName: string;
  bidderUsername: string;
  amount: number;
  createdAt: string;
}