# .migration-backup/lib/api-zod/src/generated/types/offer.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * O'price Ng marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface Offer {
  id: number;
  listingId: number;
  buyerName: string;
  buyerUsername: string;
  amount: number;
  status: string;
  /** @nullable */
  message?: string | null;
  createdAt: string;
}